Manually install the required version of the linux kernel headers.
Install the linux header packages prior to installing the vms guest
agent. We have to do this manually because we can't specify an
explicit version of the kernel headers package as a dependency in the
agent package. Note that we also do this when installing vmsfs (in the
compute recipe).