Provide additional ways to configure vms to use ceph.
We currently install vms-rados when the ceph_config key is non
nil. One can easily install ceph and use VMS without providing
ceph_config, as VMS uses a sane default.

In this change we add some optional vms sysconfig attributes for
setting rados/rbd pools and prefixes and automatically update the
memory and disk urls based on these attributes when they are
provided. With this, all a user has to do to configure vms to use
ceph is create the rados and rbd pools and provide the pool names
and the optional prefixes through the chef attributes.