Yum repositories now expect a url to key.
The new yum cookbook (3.x series) got rid of the 'yum_key' resource in
favor of using a 'key' parameter which accepts a url (similar to
apt). This is unfortunately a backwards-compatibility breaking change,
but since we have no known centos users at the moment we're modifying
the cookbook to explictly depend on the new yum cookbook.